# Fix close deadlock (variant B)