@raulk 2019-06-21
Note: This is a bit tricky. Ideally in tests I’d like to close the channel by cancelling the sub, then consume all pending events with a for range. If we let the eventbus drain events on cancel, this approach would be unfeasible. We _can_ make this the contract of the eventbus, but it does eschew legitimate use cases.