@magik6k 2019-06-21
Closing the channel by cancelling the subscription would limit us to one channel per sub, which would be mildly painful, and if we wanted to do some clever refcounting, we'd still have the deadlock problem with more than one sub on channel