@magik6k 2019-06-21
We can also enforce the cancel from all subs on the api level by removing the cancelFunc from `Subscribe`, and create new method on the eventbus:

```
type Bus interface {
  [...]
  // Cancel all related subs, close the channel
  Unsubscribe(channel interface{})
}
```