@raulk 2019-06-21
How about we re-evaluate the original proposal of the eventbus creating and owning the channel? Then all of these problems could go away. The `Subscribe()` method would be variadic and would accept N types. Yes, this implies type casts (on non-multiplexed subs) and type switches (on multiplexed subs), but the DX gets so much better. And honestly, on multiplexed subs, with the current design you'd also require type switches.