@raulk 2019-06-21
I also have the impression that this particular design will extend poorly with regards to features we ~want~ need to add, such as failing subscribers if emitters are not present.