@Kubuxu 2019-06-21
I agree that the design where the event bus owns the chan makes sense. It makes it clear who manages the channel (we can also return read-only channel to the event consumer).

```
func For(types... interface{}) SubOpt {}
func Buffer(size int) SubOpt {}

type Bus interface {
  [...]
  Subscribe(opts... SubOpt) Subscription
}
```